# artifacts/api-server/src/routes/appointments.ts
import { Router } from "express";
import { db, appointmentsTable, stylistProfilesTable, servicesTable, usersTable, paymentsTable, payoutLedgerTable } from "@workspace/db";
import { eq, and, or, desc } from "drizzle-orm";
import { randomUUID } from "crypto";
import { requireAuth } from "../lib/auth";
import { CreateAppointmentBody, UpdateAppointmentBody } from "@workspace/api-zod";
import { sendNotification, notify } from "../lib/notifications";
import { isValidSlot, isUniqueViolation } from "../lib/bookingValidation";
import { postSystemMessage } from "./messages";
import { campaignJobTotal, isCampaignJob, recordPayoutEvent, resolveCampaignCollected, splitAmount, transitionToDisputed } from "../lib/escrow";
import { param } from "../lib/params";
import { createPayoutLedgerLines } from "../lib/payout-ledger";

const router = Router();

import { ARTIST_SHARE as ARTIST_PAYOUT_PCT, PLATFORM_SHARE as PLATFORM_FEE_PCT, payoutDueAt } from "../lib/money";

function formatAppt(a: typeof appointmentsTable.$inferSelect) {
  return {
    id: a.id,
    clientId: a.clientId,
    clientName: a.clientName,
    stylistId: a.stylistId,
    stylistName: a.stylistName,
    serviceId: a.serviceId,
    serviceName: a.serviceName,
    date: a.date,
    time: a.time,
    status: a.status,
    price: a.price,
    duration: a.duration,
    notes: a.notes ?? null,
    createdAt: a.createdAt.toISOString(),
    paymentMode: a.paymentMode,
    depositAmount: a.depositAmount,
    tipAmount: a.tipAmount,
    balanceDue: a.balanceDue,
    workConfirmedByClient: a.workConfirmedByClient,
    workConfirmedByClientAt: a.workConfirmedByClientAt?.toISOString() ?? null,
    workConfirmedByArtist: a.workConfirmedByArtist,
    workConfirmedByArtistAt: a.workConfirmedByArtistAt?.toISOString() ?? null,
    payoutStatus: a.payoutStatus,
    artistPayoutAmount: a.artistPayoutAmount,
    platformFeeAmount: a.platformFeeAmount,
    campaignId: a.campaignId ?? null,
    feeMode: a.feeMode as "commission" | "brand_on_top",
  };
}

router.get("/appointments", requireAuth, async (req, res) => {
  const user = (req as any).user;
  const { status, role } = req.query as Record<string, string>;

  let appts;
  if (role === "stylist" || user.role === "stylist") {
    // appointments.stylistId holds a stylist_profiles.id, NOT a users.id.
    // Comparing it against user.id matches nothing, which is why artists were
    // getting an empty list. dashboard.ts already does this correctly.
    const [profile] = await db.select().from(stylistProfilesTable)
      .where(eq(stylistProfilesTable.userId, user.id))
      .limit(1);

    appts = profile
      ? await db.select().from(appointmentsTable).where(eq(appointmentsTable.stylistId, profile.id))
      : [];
  } else {
    appts = await db.select().from(appointmentsTable).where(eq(appointmentsTable.clientId, user.id));
  }

  if (status) {
    appts = appts.filter(a => a.status === status);
  }

  res.json(appts.map(formatAppt));
});

router.post("/appointments", requireAuth, async (req, res) => {
  const user = (req as any).user;
  const parsed = CreateAppointmentBody.safeParse(req.body);
  if (!parsed.success) { res.status(400).json({ error: "Validation error" }); return; }

  const { stylistId, serviceId, date, time, notes } = parsed.data;

  const slotCheck = isValidSlot(date, time);
  if (!slotCheck.ok) { res.status(400).json({ error: slotCheck.error }); return; }

  const [profile] = await db.select().from(stylistProfilesTable).where(eq(stylistProfilesTable.id, stylistId));
  const [service] = await db.select().from(servicesTable).where(and(
    eq(servicesTable.id, serviceId),
    eq(servicesTable.stylistId, stylistId),
  ));

  if (!profile || !service) { res.status(404).json({ error: "Stylist or service not found" }); return; }
  const [artistUser] = await db.select({ accountStatus: usersTable.accountStatus }).from(usersTable).where(eq(usersTable.id, profile.userId));
  if (artistUser?.accountStatus === "suspended") { res.status(403).json({ error: "This artist account is suspended and cannot accept bookings." }); return; }
  // Part c: block bookings with unverified artists server-side.
  if (!profile.verified) { res.status(403).json({ error: "This artist is not yet verified on Bonisa and cannot accept bookings." }); return; }

  let appt;
  try {
    [appt] = await db.insert(appointmentsTable).values({
      id: randomUUID(),
      clientId: user.id,
      clientName: user.name,
      stylistId,
      stylistName: profile.name,
      serviceId,
      serviceName: service.name,
      date,
      time,
      status: "pending",
      price: service.price,
      duration: service.duration,
      notes: notes ?? null,
    }).returning();
  } catch (err) {
    // Atomic DB constraint (appointments_stylist_slot_active_unique) rejects
    // overlapping active bookings for the same stylist/date/time — this is
    // the source of truth for conflict prevention, not a pre-check, so it's
    // race-condition safe under concurrent requests.
    if (isUniqueViolation(err)) {
      res.status(409).json({ error: "This time slot was just booked by someone else. Please choose another time." });
      return;
    }
    throw err;
  }

  // Notify stylist of new booking (non-fatal)
  try {
            const [stylistUser] = await db.select().from(usersTable).where(eq(usersTable.id, profile.userId));
    await sendNotification(stylistUser?.phone, "booking.created", {
      clientName: user.name,
      serviceName: service.name,
      date,
      time,
    });
  } catch { /* non-fatal */ }

  res.status(201).json(formatAppt(appt));
});

router.get("/appointments/:appointmentId", requireAuth, async (req, res) => {
  const user = (req as any).user;
  const appointmentId = param(req.params.appointmentId);
  const [appt] = await db.select().from(appointmentsTable)
    .where(eq(appointmentsTable.id, appointmentId));
  if (!appt) { res.status(404).json({ error: "Not found" }); return; }
  const [callerProfile] = await db.select().from(stylistProfilesTable)
    .where(and(
      eq(stylistProfilesTable.id, appt.stylistId),
      eq(stylistProfilesTable.userId, user.id),
    ));
  if (appt.clientId !== user.id && !callerProfile) {
    res.status(403).json({ error: "You are not a participant in this booking" });
    return;
  }
  res.json(formatAppt(appt));
});

router.patch("/appointments/:appointmentId", requireAuth, async (req, res) => {
  const user = (req as any).user;
  const appointmentId = param(req.params.appointmentId);
  const parsed = UpdateAppointmentBody.safeParse(req.body);
  if (!parsed.success) { res.status(400).json({ error: "Validation error" }); return; }
  const data = parsed.data;

  // Fetch the appointment before updating so we have context for notifications
  const [before] = await db.select().from(appointmentsTable).where(eq(appointmentsTable.id, appointmentId));
  if (!before) { res.status(404).json({ error: "Not found" }); return; }

  const isClient = before.clientId === user.id;
  const [callerProfile] = await db.select().from(stylistProfilesTable)
    .where(and(
      eq(stylistProfilesTable.id, before.stylistId),
      eq(stylistProfilesTable.userId, user.id),
    ));
  const isArtist = !!callerProfile;
  if (!isClient && !isArtist) {
    res.status(403).json({ error: "You are not a participant in this booking" });
    return;
  }

  if (isArtist) {
    if (data.date !== undefined || data.time !== undefined || data.notes !== undefined) {
      res.status(403).json({ error: "Only the client can edit booking details" });
      return;
    }
    if (data.status !== "confirmed" && data.status !== "declined") {
      res.status(403).json({ error: "Artists can only confirm or decline booking requests" });
      return;
    }
    if (before.status !== "pending") {
      res.status(409).json({ error: "Only pending bookings can be confirmed or declined" });
      return;
    }
  } else {
    if (data.status && data.status !== "cancelled") {
      res.status(403).json({ error: "Clients can only cancel bookings" });
      return;
    }
    if ((data.date !== undefined || data.time !== undefined || data.notes !== undefined) && before.status !== "pending") {
      res.status(409).json({ error: "Only pending booking details can be edited" });
      return;
    }
    if (data.status === "cancelled" && !["pending", "confirmed"].includes(before.status)) {
      res.status(409).json({ error: "This booking can no longer be cancelled" });
      return;
    }
  }

  const [appt] = await db.update(appointmentsTable).set({
    ...(data.status && { status: data.status }),
    ...(data.date && { date: data.date }),
    ...(data.time && { time: data.time }),
    ...(data.notes !== undefined && { notes: data.notes }),
  }).where(eq(appointmentsTable.id, appointmentId)).returning();
  if (!appt) { res.status(404).json({ error: "Not found" }); return; }

  res.json(formatAppt(appt));

  // Fire notifications asynchronously after responding — non-fatal
  if (data.status && data.status !== before.status) {
    setImmediate(async () => {
      try {
        const [clientUser] = await db.select().from(usersTable).where(eq(usersTable.id, appt.clientId));
        const sharedData = {
          stylistName: appt.stylistName,
          clientName: appt.clientName,
          serviceName: appt.serviceName,
          date: appt.date,
          time: appt.time,
        };

        if (data.status === "confirmed") {
          // Notify client
          await sendNotification(clientUser?.phone, "booking.confirmed", sharedData);
          // Notify stylist too — both sides should know a booking is locked in
          const [confirmedProfile] = await db.select().from(stylistProfilesTable).where(eq(stylistProfilesTable.id, appt.stylistId));
          if (confirmedProfile) {
            const [confirmedStylistUser] = await db.select().from(usersTable).where(eq(usersTable.id, confirmedProfile.userId));
            await sendNotification(confirmedStylistUser?.phone, "booking.confirmed.stylist", sharedData);
          }
          await postSystemMessage(
            appt.clientId, appt.stylistId,
            `✅ Booking confirmed — ${appt.serviceName} on ${appt.date} at ${appt.time}.`
          );
        } else if (data.status === "declined" || data.status === "cancelled") {
          await sendNotification(clientUser?.phone, "booking.declined", sharedData);
          await postSystemMessage(
            appt.clientId, appt.stylistId,
            `❌ Booking ${data.status} — ${appt.serviceName} on ${appt.date} at ${appt.time}.`
          );
        } else if (data.status === "completed") {
          // Notify client
          await sendNotification(clientUser?.phone, "booking.completed", sharedData);
          // Notify stylist too
          const [profile] = await db.select().from(stylistProfilesTable).where(eq(stylistProfilesTable.id, appt.stylistId));
          if (profile) {
            const [stylistUser] = await db.select().from(usersTable).where(eq(usersTable.id, profile.userId));
            await sendNotification(stylistUser?.phone, "booking.completed", sharedData);
          }
          await postSystemMessage(
            appt.clientId, appt.stylistId,
            `🌟 Appointment complete — ${appt.serviceName}. Please leave a review!`
          );
        }
      } catch { /* non-fatal */ }
    });
  }
});

// ── Dual Work Confirmation + Payout Split ────────────────────────────────────
/**
 * POST /appointments/:appointmentId/confirm-work
 * Body: { dispute?: boolean }
 *
 * Either the client or the artist calls this to confirm work was done.
 * - If dispute=true  → payoutStatus = 'disputed', frozen for manual review
 * - If both confirm  → payoutStatus = 'released', 82/18 split computed,
 *                      appointment status set to 'completed'
 */
router.post("/appointments/:appointmentId/confirm-work", requireAuth, async (req, res) => {
  const user = (req as any).user;
  const appointmentId = String(req.params.appointmentId);
  const dispute = req.body?.dispute === true;

  const [appt] = await db.select().from(appointmentsTable)
    .where(eq(appointmentsTable.id, appointmentId));
  if (!appt) { res.status(404).json({ error: "Appointment not found" }); return; }

  // Determine caller role
  const isClient = appt.clientId === user.id;
  const stylistRows = await db.select().from(stylistProfilesTable)
    .where(and(eq(stylistProfilesTable.id, appt.stylistId), eq(stylistProfilesTable.userId, user.id)));
  const isArtist = stylistRows.length > 0;

  if (!isClient && !isArtist) {
    res.status(403).json({ error: "Not authorised to confirm this appointment" }); return;
  }

  // State machine: only paid, confirmed bookings enter the completion flow.
  if (appt.status !== "confirmed" && appt.status !== "completed") {
    res.status(409).json({ error: "This booking is not in a confirmable state." }); return;
  }
  const [paymentRow] = await db.select().from(paymentsTable)
    .where(eq(paymentsTable.appointmentId, appointmentId))
    .orderBy(desc(paymentsTable.createdAt))
    .limit(1);
  const isPaid = (paymentRow && paymentRow.status === "succeeded") || !!appt.stripePaymentIntentId;
  if (!isPaid) {
    res.status(409).json({ error: "No verified payment for this booking — nothing is held in escrow yet." }); return;
  }

  if (dispute) {
    // Disputes can only freeze funds still in escrow. A released payout is
    // final here; post-release complaints go through support, not the ledger.
    const result = await db.transaction(tx => transitionToDisputed(tx, appt, user.id,
      `Issue reported by ${isClient ? "client" : "artist"} — funds frozen in escrow pending review`));
    const updated = result.appointment;
    if (!updated) {
      const state = appt.payoutStatus === "released"
        ? "Payout was already released. Please contact support to raise a complaint."
        : "This booking is already under dispute.";
      res.status(409).json({ error: state });
      return;
    }
    try {
      await postSystemMessage(
        appt.clientId, appt.stylistId,
        `⚠️ A dispute has been raised on the ${appt.serviceName} appointment (${appt.date}). The payout is frozen — Bonisa will reach out to both parties to resolve it.`
      );
    } catch { /* non-fatal */ }
    res.json(formatAppt(updated));
    return;
  }

  if (appt.payoutStatus === "disputed") {
    res.status(409).json({ error: "This appointment is under dispute. Please contact support." }); return;
  }
  if (appt.payoutStatus === "released") {
    res.status(409).json({ error: "Payout already released for this appointment." }); return;
  }

  // On a campaign the brand pays in two parts, so it cannot say "work done" until it has paid in full.
  // That keeps a half-paid campaign from ever being marked complete. The artist can confirm early;
  // her confirmation simply waits for the brand.
  if (isClient && isCampaignJob(appt)) {
    const collectedSoFar = await resolveCampaignCollected(db, appointmentId);
    if (Math.round(collectedSoFar * 100) < Math.round(campaignJobTotal(appt) * 100)) {
      res.status(409).json({ error: "Please pay the remaining campaign balance first. Your artists are paid once the campaign is paid in full." });
      return;
    }
  }

  // Record this party's confirmation with a conditional update so concurrent
  // retries can't double-confirm: the WHERE clause only matches while the flag
  // is still false and funds are still held.
  const now = new Date();
  const confirmedRows = await db.update(appointmentsTable)
    .set(isClient
      ? { workConfirmedByClient: true, workConfirmedByClientAt: now }
      : { workConfirmedByArtist: true, workConfirmedByArtistAt: now })
    .where(and(
      eq(appointmentsTable.id, appointmentId),
      eq(appointmentsTable.payoutStatus, "held"),
      isClient
        ? eq(appointmentsTable.workConfirmedByClient, false)
        : eq(appointmentsTable.workConfirmedByArtist, false),
    ))
    .returning();

  if (confirmedRows.length > 0) {
    await recordPayoutEvent({
      appointmentId, type: isClient ? "client_confirmed" : "artist_confirmed",
      actorUserId: user.id,
      note: isClient ? "Client confirmed the appointment was received" : "Artist confirmed the appointment was completed",
    });
  }

  // Attempt release whenever both flags are set and funds are still held.
  // Running this even when the confirmation above was a no-op means a booking
  // that previously confirmed both sides but failed to release gets retried
  // here instead of being stuck. The conditional WHERE makes release atomic —
  // exactly one caller wins; everyone else matches zero rows.
  // A campaign job is paid in two parts, so what has been collected is the sum
  // of every payment, and the artist is only paid out once the brand has paid in full.
  const campaignJob = isCampaignJob(appt);
  const actualCollected = campaignJob
    ? await resolveCampaignCollected(db, appointmentId)
    : paymentRow?.amount
      ?? (appt.paymentMode === "deposit" && appt.depositAmount > 0
        ? appt.depositAmount + appt.tipAmount
        : appt.price + appt.tipAmount);
  if (campaignJob && Math.round(actualCollected * 100) < Math.round(campaignJobTotal(appt) * 100)) {
    // The artist has confirmed, but the brand has not paid in full yet. Her confirmation is saved
    // and the payout waits, exactly like waiting for the other side.
    const [current] = await db.select().from(appointmentsTable).where(eq(appointmentsTable.id, appointmentId));
    res.json(formatAppt(current ?? appt));
    return;
  }
  const { artistShare: artistPayout, platformShare: platformFee } = splitAmount(actualCollected, appt.feeMode);

  const releasedRows = await db.transaction(async (tx) => {
    const rows = await tx.update(appointmentsTable)
      .set({
        payoutStatus: "released",
        artistPayoutAmount: artistPayout,
        platformFeeAmount: platformFee,
        status: "completed",
      })
      .where(and(
        eq(appointmentsTable.id, appointmentId),
        eq(appointmentsTable.payoutStatus, "held"),
        eq(appointmentsTable.workConfirmedByClient, true),
        eq(appointmentsTable.workConfirmedByArtist, true),
      ))
      .returning();
    if (rows.length > 0) await createPayoutLedgerLines(tx, rows[0], payoutDueAt());
    return rows;
  });

  if (releasedRows.length > 0) {
    await recordPayoutEvent({
      appointmentId, type: "released", actorUserId: user.id,
      amount: actualCollected, artistShare: artistPayout, platformShare: platformFee,
      note: "Both parties confirmed — escrow released",
    });
    try {
      await postSystemMessage(
        appt.clientId, appt.stylistId,
        appt.isTeamBooking
          ? `Both parties confirmed the appointment. R${artistPayout.toFixed(2)} allocated across the confirmed team artists (Bonisa fee: R${platformFee.toFixed(2)}).`
          : `Both parties confirmed the appointment. R${artistPayout.toFixed(2)} released to ${appt.stylistName} (Bonisa fee: R${platformFee.toFixed(2)}).`
      );
    } catch { /* non-fatal */ }
    // Notify each paid artist of only their own allocation — email + WhatsApp, non-fatal.
    setImmediate(async () => {
      try {
        const ledgerLines = await db.select({
          artistProfileId: payoutLedgerTable.artistProfileId,
          amount: payoutLedgerTable.netAmount,
        }).from(payoutLedgerTable).where(eq(payoutLedgerTable.appointmentId, appointmentId));
        const dueDate = payoutDueAt().toLocaleDateString("en-ZA");
        for (const line of ledgerLines) {
          const [artistProfile] = await db.select().from(stylistProfilesTable)
            .where(eq(stylistProfilesTable.id, line.artistProfileId));
          if (!artistProfile) continue;
          const [artistUser] = await db.select().from(usersTable)
            .where(eq(usersTable.id, artistProfile.userId));
          if (!artistUser) continue;
          await notify(
            { phone: artistUser.phone, email: artistUser.email, name: artistUser.name },
            "payout.released",
            { artistName: artistProfile.name, serviceName: appt.serviceName, amount: line.amount, payoutDueAt: dueDate },
          );
        }
      } catch { /* non-fatal */ }
    });
    res.json(formatAppt(releasedRows[0]));
    return;
  }

  // Not released — return the freshest state
  const [current] = await db.select().from(appointmentsTable)
    .where(eq(appointmentsTable.id, appointmentId));
  res.json(formatAppt(current ?? appt));
});

export default router;