# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: GlamNet API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /appointments:
    get:
      operationId: listAppointments
      tags: [appointments]
      summary: List appointments for current user
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, completed, cancelled, declined]
        - name: role
          in: query
          schema:
            type: string
            enum: [client, stylist]
      responses:
        "200":
          description: Appointments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Appointment"

    post:
      operationId: createAppointment
      tags: [appointments]
      summary: Book an appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
  /appointments/{appointmentId}:
    get:
      operationId: getAppointment
      tags: [appointments]
      summary: Get appointment details
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"

    patch:
      operationId: updateAppointment
      tags: [appointments]
      summary: Update appointment status
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Appointment"
...
components:
  schemas:
...
    Appointment:
      type: object
      required: [id, clientId, stylistId, serviceId, serviceName, date, time, status, price, duration, createdAt]
      properties:
        campaignId:
          type: string
          nullable: true
        feeMode:
          type: string
          enum: [commission, brand_on_top]
        id:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        stylistId:
          type: string
        stylistName:
          type: string
        serviceId:
          type: string
        serviceName:
          type: string
        date:
          type: string
        time:
          type: string
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled, declined]
        price:
          type: number
        duration:
          type: integer
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    AppointmentInput:
      type: object
      required: [stylistId, serviceId, date, time]
      properties:
        stylistId:
          type: string
        serviceId:
          type: string
        date:
          type: string
        time:
          type: string
        notes:
          type: string
    AppointmentUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, completed, cancelled, declined]
        date:
          type: string
        time:
          type: string
        notes:
          type: string
...